mogami: don't aggressively disable wifi configurations
Wifi networks get blacklisted for a long time when the initial join
or requesting an address via DHCP fails. They should still be tried
when the device has no WiFi connection at all, but this doesn't seem
to work correctly with older wifi-drivers. Set the blacklist-time
to 0 to disable this "feature".

Change-Id: Ia216738dd598fc2f16119a5ed5a3252c1cae31c5